+++ trailmet/algorithms/quantize/q_utils.py
@@ -1,6 +1,4 @@
 import torch
-import torch.nn as nn
-import torch.nn.init as init
 import time
 
 class AverageMeter(object):
@@ -27,7 +25,7 @@ class AverageMeter(object):
         return fmtstr.format(**self.__dict__)
 
 class ProgressMeter(object):
-    """Helper for printing calibration status"""
+    """Helper for displaying progress"""
     def __init__(self, num_batches, meters, prefix=""):
         self.batch_fmtstr = self._get_batch_fmtstr(num_batches)
         self.meters = meters
@@ -44,35 +42,8 @@ class ProgressMeter(object):
         return '[' + fmt + '/' + fmt.format(num_batches) + ']'
 
 
-class StraightThrough(nn.Module):
-    def __int__(self):
-        super().__init__()
-
-    def forward(self, input):
-        return input
-
-
-def round_ste(x: torch.Tensor):
-    """
-    Implement Straight-Through Estimator for rounding operation.
-    """
-    return (x.round() - x).detach() + x
-
-
-def lp_loss(pred, tgt, p=2.0, reduction='none'):
-    """
-    loss function measured in L_p Norm
-    """
-    if reduction == 'none':
-        return (pred-tgt).abs().pow(p).sum(1).mean()
-    else:
-        return (pred-tgt).abs().pow(p).mean()
-
-
 def accuracy(output, target, topk=(1,)):
-    """
-    Computes the accuracy over the k top predictions for the specified values of k
-    """
+    """Computes the accuracy over the k top predictions for the specified values of k"""
     with torch.no_grad():
         maxk = max(topk)
         batch_size = target.size(0)
@@ -129,87 +100,7 @@ def validate_model(val_loader, model, device=None, print_freq=100):
 
     return top1.avg
 
-
-def _fold_bn(conv_module, bn_module):
-    w = conv_module.weight.data
-    y_mean = bn_module.running_mean
-    y_var = bn_module.running_var
-    safe_std = torch.sqrt(y_var + bn_module.eps)
-    w_view = (conv_module.out_channels, 1, 1, 1)
-    if bn_module.affine:
-        weight = w * (bn_module.weight / safe_std).view(w_view)
-        beta = bn_module.bias - bn_module.weight * y_mean / safe_std
-        if conv_module.bias is not None:
-            bias = bn_module.weight * conv_module.bias / safe_std + beta
-        else:
-            bias = beta
-    else:
-        weight = w / safe_std.view(w_view)
-        beta = -y_mean / safe_std
-        if conv_module.bias is not None:
-            bias = conv_module.bias / safe_std + beta
-        else:
-            bias = beta
-    return weight, bias
-
-
-def fold_bn_into_conv(conv_module, bn_module):
-    w, b = _fold_bn(conv_module, bn_module)
-    if conv_module.bias is None:
-        conv_module.bias = nn.Parameter(b)
-    else:
-        conv_module.bias.data = b
-    conv_module.weight.data = w
-    # set bn running stats
-    bn_module.running_mean = bn_module.bias.data
-    bn_module.running_var = bn_module.weight.data ** 2
-
-
-def reset_bn(module: nn.BatchNorm2d):
-    if module.track_running_stats:
-        module.running_mean.zero_()
-        module.running_var.fill_(1-module.eps)
-        # we do not reset numer of tracked batches here
-        # self.num_batches_tracked.zero_()
-    if module.affine:
-        init.ones_(module.weight)
-        init.zeros_(module.bias)
-
-
-def is_bn(m):
-    return isinstance(m, nn.BatchNorm2d) or isinstance(m, nn.BatchNorm1d)
-
-
-def is_absorbing(m):
-    return (isinstance(m, nn.Conv2d)) or isinstance(m, nn.Linear)
-
-
-def search_fold_and_remove_bn(model):
-    model.eval()
-    prev = None
-    for n, m in model.named_children():
-        if is_bn(m) and is_absorbing(prev):
-            fold_bn_into_conv(prev, m)
-            # set the bn module to straight through
-            setattr(model, n, StraightThrough())
-        elif is_absorbing(m):
-            prev = m
-        else:
-            prev = search_fold_and_remove_bn(m)
-    return prev
-
-
-def search_fold_and_reset_bn(model):
-    model.eval()
-    prev = None
-    for n, m in model.named_children():
-        if is_bn(m) and is_absorbing(prev):
-            fold_bn_into_conv(prev, m)
-            # reset_bn(m)
-        else:
-            search_fold_and_reset_bn(m)
-        prev = m
-
+# To-Do : Modify for balanced subsampling 
 
 def get_train_samples(train_loader, num_samples):
     train_data = []
